requirements.txt
python-telegram-bot==13.12
environs==9.5.0
phonenumbers==8.13.6